make deployer helm tests extendable
