+++ test/deployer/testdata/base-gateway-out.yaml
