+++ test/deployer/testdata/omit-default-security-context-out.yaml
