+++ test/deployer/testdata/omit-default-security-context.yaml
