+++ test/deployer/testutils.go
@@ -1,0 +1,85 @@
+package deployer
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/upstreams/http/v3"
+	"istio.io/istio/pkg/kube/krt/krttest"
+	"istio.io/istio/pkg/test"
+	corev1 "k8s.io/api/core/v1"
+	rbacv1 "k8s.io/api/rbac/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+	inf "sigs.k8s.io/gateway-api-inference-extension/api/v1"
+	api "sigs.k8s.io/gateway-api/apis/v1"
+	apixv1a1 "sigs.k8s.io/gateway-api/apisx/v1alpha1"
+
+	"github.com/kgateway-dev/kgateway/v2/api/settings"
+	"github.com/kgateway-dev/kgateway/v2/internal/kgateway/krtcollections"
+	"github.com/kgateway-dev/kgateway/v2/internal/kgateway/wellknown"
+	sdk "github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk"
+	"github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk/collections"
+	"github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk/krtutil"
+	"github.com/kgateway-dev/kgateway/v2/pkg/schemes"
+)
+
+func NewCommonCols(t test.Failer, initObjs ...client.Object) *collections.CommonCollections {
+	ctx := context.Background()
+	var anys []any
+	for _, obj := range initObjs {
+		anys = append(anys, obj)
+	}
+	mock := krttest.NewMock(t, anys)
+
+	policies := krtcollections.NewPolicyIndex(krtutil.KrtOptions{}, sdk.ContributesPolicies{}, settings.Settings{})
+	kubeRawGateways := krttest.GetMockCollection[*api.Gateway](mock)
+	kubeRawListenerSets := krttest.GetMockCollection[*apixv1a1.XListenerSet](mock)
+	gatewayClasses := krttest.GetMockCollection[*api.GatewayClass](mock)
+	nsCol := krtcollections.NewNamespaceCollectionFromCol(ctx, krttest.GetMockCollection[*corev1.Namespace](mock), krtutil.KrtOptions{})
+
+	krtopts := krtutil.NewKrtOptions(ctx.Done(), nil)
+	gateways := krtcollections.NewGatewayIndex(krtopts, wellknown.DefaultGatewayControllerName, policies, kubeRawGateways, kubeRawListenerSets, gatewayClasses, nsCol)
+
+	commonCols := &collections.CommonCollections{
+		GatewayIndex: gateways,
+	}
+
+	for !kubeRawGateways.HasSynced() || !kubeRawListenerSets.HasSynced() || !gatewayClasses.HasSynced() {
+		time.Sleep(time.Second / 10)
+	}
+
+	gateways.Gateways.WaitUntilSynced(ctx.Done())
+	return commonCols
+}
+
+// initialize a fake controller-runtime client with the given list of objects
+func NewFakeClientWithObjs(objs ...client.Object) client.Client {
+	scheme := schemes.GatewayScheme()
+	return NewFakeClientWithObjsWithScheme(scheme, objs...)
+}
+
+func NewFakeClientWithObjsWithScheme(scheme *runtime.Scheme, objs ...client.Object) client.Client {
+	// Ensure the rbac types are registered.
+	if err := rbacv1.AddToScheme(scheme); err != nil {
+		panic(fmt.Sprintf("failed to add rbacv1 scheme: %v", err))
+	}
+
+	// Check if any object is an InferencePool, and add its scheme if needed.
+	for _, obj := range objs {
+		gvk := obj.GetObjectKind().GroupVersionKind()
+		if gvk.Kind == wellknown.InferencePoolKind {
+			if err := inf.Install(scheme); err != nil {
+				panic(fmt.Sprintf("failed to add InferenceExtension scheme: %v", err))
+			}
+			break
+		}
+	}
+
+	return fake.NewClientBuilder().
+		WithScheme(scheme).
+		WithObjects(objs...).
+		Build()
+}
